优化部分单元测试代码，以规避潜在的 bug
com.dwarfeng.rbacds.impl.service.FilterSupportMaintainServiceImplTest
com.dwarfeng.rbacds.impl.service.RoleMaintainServiceImplTest
com.dwarfeng.rbacds.impl.service.UserMaintainServiceImplTest

+++ rbac-distributed-service-impl/src/test/java/com/dwarfeng/rbacds/impl/service/FilterSupportMaintainServiceImplTest.java
@@ -14,6 +14,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 
@@ -59,6 +60,9 @@ public class FilterSupportMaintainServiceImplTest {
             }
         } finally {
             for (FilterSupport filterSupport : filterSupports) {
+                if (Objects.isNull(filterSupport.getKey())) {
+                    continue;
+                }
                 service.deleteIfExists(filterSupport.getKey());
             }
         }
